Rasterization MVP

+++ src/clear_rendertarget.h
@@ -31,17 +31,17 @@ namespace cg
 	class ClearRenderTarget
 	{
 	public:
-		ClearRenderTarget(unsigned short width, unsigned short height);
+		ClearRenderTarget(unsigned width, unsigned height);
 		virtual ~ClearRenderTarget();
 
 		void Clear();
 		void Save(std::string filename) const;
 
 	protected:
-		void SetPixel(unsigned short x, unsigned short y, color color);
+		void SetPixel(unsigned x, unsigned y, color color);
 
-		unsigned short height;
-		unsigned short width;
+		unsigned height;
+		unsigned width;
 
 		std::vector<color> frame_buffer;
 	};
