+++ src/draw_line.h
@@ -12,7 +12,7 @@ namespace cg
 		LineDrawing(unsigned short width, unsigned short height);
 		virtual ~LineDrawing();
 
-		void DrawLine(unsigned short x_begin, unsigned short y_begin, unsigned short x_end, unsigned short y_end, color color);
+		void DrawLine(unsigned x_begin, unsigned y_begin, unsigned x_end, unsigned y_end, color color);
 
 		void DrawScene();
